+++ src/handler/config.ts
@@ -1,0 +1,8 @@
+import config from '~/secret/config.json';
+
+export const token = config.token;
+export const prefix = config.prefix;
+export const apiKey = config['google-api-key'];
+export const ytChannelId = config['yt-channel-id'];
+
+export default config;
